pyproject.toml
[tool.poetry]
name = "mosaic"
version = "0.1.0"
description = ""
authors = ["Loic Coyle <loic.coyle@hotmail.fr>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
Pillow = "^8.2.0"
numpy = "^1.20.2"
tqdm = "^4.60.0"
matplotlib = "^3.4.1"
colour-science = "^0.3.16"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
mosaic = 'mosaic.__main__:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
